Further updates to Virtual_HID and testcases
Virtual_HID has been refactored quite a bit in order to keep
it in accordance with PEP 8. A fully functional VMouse is now
available, although further changes are likely to occur. It
has been fully tested (thus far). VKeyboard still needs quite
a bit of work, which is next on the agenda. Also removed some
environment files that are not necessary for the project
itself.