Addition of weekend work.
Made some changes around the place to do with imports. Still trying to get
the hang of git, so removing the .pyc files, .project and .pydevproject.
Hopefully this time will work. Looking into logging and this should be
implemented soon. More work to be done on VKeyboard such as parsing, unittests
and actual input of n-keys (n=2, 3, ...).