Simple UI added. Installation & Usage pdf added.
Started on a simple UI. Still a lot of work to be done,
but scripts can be run through it, saved and editted.

Added an Installation & Usage guide in the Doc folder.
This will need to be updated frequently.

As for the UI, there is a mockup in the Doc folder. Pretty.
No bugfixes this time, many unittests still need to be added.